refactored dataset method to be included in base mapper
