+++ abis_mapping/templates/survey_metadata/mapping.py
@@ -17,10 +17,6 @@ import rdflib
 from typing import Optional, Iterator, Any
 
 
-# Default Dataset Metadata
-DATASET_DEFAULT_NAME = "Example Systematic Survey Metadata Dataset"
-DATASET_DEFAULT_DESCRIPTION = "Example Systematic Survey Metadata Dataset by Gaia Resources"
-
 # Constants / shortcuts
 a = rdflib.RDF.type
 
@@ -28,6 +24,10 @@ a = rdflib.RDF.type
 class SurveyMetadataMapper(base.mapper.ABISMapper):
     template_id = "survey_metadata.csv"
     instructions_file = "instructions.pdf"
+
+    # Default Dataset Metadata
+    DATASET_DEFAULT_NAME = "Example Systematic Survey Metadata Dataset"
+    DATASET_DEFAULT_DESCRIPTION = "Example Systematic Survey Metadata Dataset by Gaia Resources"
 
     def apply_validation(
             self,
@@ -115,7 +115,7 @@ class SurveyMetadataMapper(base.mapper.ABISMapper):
         # Check if Dataset IRI Supplied
         if not dataset_iri:
             # Create Dataset IRI
-            dataset_iri = utils.rdf.uri(f"dataset/{DATASET_DEFAULT_NAME}", base_iri)
+            dataset_iri = utils.rdf.uri(f"dataset/{self.DATASET_DEFAULT_NAME}", base_iri)
 
             # Add the default dataset
             self.add_default_dataset(
@@ -137,22 +137,6 @@ class SurveyMetadataMapper(base.mapper.ABISMapper):
 
             yield graph
 
-    def add_default_dataset(
-        self,
-        uri: rdflib.URIRef,
-        graph: rdflib.Graph,
-    ) -> None:
-        """Adds Default Example Dataset to the Graph
-
-        Args:
-            uri (rdflib.URIRef): IRI of the dataset.
-            graph (rdflib.Graph): Graph to add to.
-        """
-        # Add Default Dataset to Graph
-        graph.add((uri, a, utils.namespaces.TERN.RDFDataset))
-        graph.add((uri, rdflib.DCTERMS.title, rdflib.Literal(DATASET_DEFAULT_NAME)))
-        graph.add((uri, rdflib.DCTERMS.description, rdflib.Literal(DATASET_DEFAULT_DESCRIPTION)))
-        graph.add((uri, rdflib.DCTERMS.issued, utils.rdf.toTimestamp(datetime.date.today())))
 
     def apply_mapping_row(
         self,
